requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
omegaconf>=2.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
plotly>=5.15.0

# PyTorch ecosystem
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
einops>=0.6.0
albumentations>=1.3.0

# Text-to-Image specific
transformers>=4.30.0
diffusers>=0.20.0
accelerate>=0.20.0
safetensors>=0.3.0
xformers>=0.0.20; sys_platform != "darwin" or platform_machine != "arm64"

# Training and evaluation
pytorch-lightning>=2.0.0
torchmetrics>=1.0.0
lpips>=0.1.4
clean-fid>=0.1.35
wandb>=0.15.0
tensorboard>=2.13.0

# Development and testing
black>=23.0.0
ruff>=0.0.280
pytest>=7.4.0
pre-commit>=3.3.0

# Demo and serving
streamlit>=1.25.0
fastapi>=0.100.0
uvicorn>=0.23.0
gradio>=3.40.0

# Image processing
pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0
